Rework of feed loader to be fully asynchronous.
Fixes #27.
Fixes #8.

Now that each HTTP request is run in its own thread, there is no need
for the feed loader BHandler. We can use an asynchronous listener to get
notified of changes in a safe and serialized way in the main application
thread.

The listener will cancel previous requests if a new one happens, and
make sure to drop any old results.